Create an environment.yml file for ibdreduce: ----
Proposed change adds an environment.yml file titled
"ibdreduce_environment.yml" file to the main directory of carvaIBD. This
change is proposed because ibdreduce.py is dependent upon dependencies
that are not in the standard library such as zstandard, numpy, and
scipy so the user needs to install them. The yaml file is for convience
so the user can just run "conda env create -f ibdreduce_environment.yml"
to install all of the necessary dependencies.

1. This file will create a conda environment called ibdreduce. The user
can activate this using the command: "conda activate ibdreduce".

2. This file was create from the below lab's server linux environment
which is: Ubuntu 20.04.5 LTS (GNU/Linux 5.4.0-135-generic x86_64). I am
not certain if it would work on other OSes but I am assuming this will
be running on linux servers

3. I created this with a python=3.10 dependency. I have been running
ibdreduce with python 3.10 so the ibdreduce code runs fine in python
3.10. But the dependencies installed by the yaml file are resolved for
python 3.10 so they will probably not work with python 3.6 which is the
current minimum version required by ibdreduce. I could change this if
necessary.